// src/lib/types.ts
export interface Listing {
  name: string;
  category: string;
  description: string;
  address: string;
  phone: string;
  whatsapp: string;
  businessHours: string;
  weeklyOff: string;
  topProducts: string;
  facebook: string;
  instagram: string;
  website: string;
  latitude: string;
  longitude: string;
  imageUrl: string;
  slug: string;
  originalUrl: string;
  dateAdded: string;
  status: string;
}